maint: re-org the pixeldata module
